Add files via upload
The GridGUI.py file is the main one that we should use for the project. The CreatingGrid.py file makes the test files for when we are ready to start testing. The ReadingTextFile.py file is just a separate file that reads a given text file and outputs a grid.